+++ Classes/Synchronisation/AbstractSynchroniser.php
@@ -15,6 +15,7 @@ use Brotkrueml\JobRouterConnector\RestClient\RestClientFactory;
 use Brotkrueml\JobRouterData\Domain\Model\Table;
 use Brotkrueml\JobRouterData\Domain\Repository\JobRouter\JobDataRepository;
 use Brotkrueml\JobRouterData\Domain\Repository\TableRepository;
+use Psr\EventDispatcher\EventDispatcherInterface;
 use Psr\Log\LoggerAwareInterface;
 use Psr\Log\LoggerAwareTrait;
 use TYPO3\CMS\Core\Database\ConnectionPool;
@@ -32,6 +33,11 @@ abstract class AbstractSynchroniser implements LoggerAwareInterface
     protected $connectionPool;
 
     /**
+     * @var EventDispatcherInterface
+     */
+    protected $eventDispatcher;
+
+    /**
      * @var RestClientFactory
      */
     private $restClientFactory;
@@ -43,10 +49,12 @@ abstract class AbstractSynchroniser implements LoggerAwareInterface
 
     public function __construct(
         ConnectionPool $connectionPool,
+        EventDispatcherInterface $eventDispatcher,
         RestClientFactory $restClientFactory,
         TableRepository $tableRepository
     ) {
         $this->connectionPool = $connectionPool;
+        $this->eventDispatcher = $eventDispatcher;
         $this->restClientFactory = $restClientFactory;
         $this->tableRepository = $tableRepository;
     }