+++ Tests/Unit/Event/ModifyDatasetOnSynchronisationEventTest.php
@@ -1,0 +1,129 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the "jobrouter_data" extension for TYPO3 CMS.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ */
+
+namespace Brotkrueml\JobRouterData\Tests\Unit\Event;
+
+use Brotkrueml\JobRouterData\Domain\Model\Table;
+use Brotkrueml\JobRouterData\Event\ModifyDatasetOnSynchronisationEvent;
+use Brotkrueml\JobRouterData\Exception\ModifyDatasetException;
+use PHPUnit\Framework\TestCase;
+
+final class ModifyDatasetOnSynchronisationEventTest extends TestCase
+{
+    /**
+     * @var Table
+     */
+    private $table;
+    /**
+     * @var array
+     */
+    private $dataset;
+    /**
+     * @var ModifyDatasetOnSynchronisationEvent
+     */
+    private $subject;
+
+    protected function setUp(): void
+    {
+        $this->table = new Table();
+        $this->table->setHandle('some handle');
+
+        $this->dataset = [
+            'jrid' => 42,
+            'field1' => 'content of field 1',
+            'field2' => 'content of field 2',
+        ];
+
+        $this->subject = new ModifyDatasetOnSynchronisationEvent($this->table, $this->dataset);
+    }
+
+    /**
+     * @test
+     */
+    public function getTable(): void
+    {
+        self::assertSame($this->table, $this->subject->getTable());
+    }
+
+    /**
+     * @test
+     */
+    public function getDataset(): void
+    {
+        self::assertSame($this->dataset, $this->subject->getDataset());
+    }
+
+    /**
+     * @test
+     */
+    public function isRejectedIsFalseAfterInstantiationOfEvent(): void
+    {
+        self::assertFalse($this->subject->isRejected());
+    }
+
+    /**
+     * @test
+     */
+    public function isRejectedIsTrueAfterSetRejectedIsCalled(): void
+    {
+        $this->subject->setRejected();
+
+        self::assertTrue($this->subject->isRejected());
+    }
+
+    /**
+     * @test
+     */
+    public function getAndSetDataset(): void
+    {
+        $dataset = [
+            'jrid' => 42,
+            'field1' => 'new content of field 1',
+            'field2' => 'new content of field 2',
+        ];
+        $this->subject->setDataset($dataset);
+
+        self::assertSame($dataset, $this->subject->getDataset());
+    }
+
+    /**
+     * @test
+     */
+    public function setDatasetThrowsExceptionWhenDatasetKeysAreRemoved(): void
+    {
+        $this->expectException(ModifyDatasetException::class);
+        $this->expectExceptionCode(1639132693);
+        $this->expectExceptionMessage('Given dataset keys "jrid, field1" differ from original dataset keys "jrid, field1, field2" when modfying dataset for table with handle "some handle"');
+
+        $dataset = [
+            'jrid' => 42,
+            'field1' => 'new content of field 1',
+        ];
+        $this->subject->setDataset($dataset);
+    }
+
+    /**
+     * @test
+     */
+    public function setDatasetThrowsExceptionWhenJridIsChangedFromDataset(): void
+    {
+        $this->expectException(ModifyDatasetException::class);
+        $this->expectExceptionCode(1639132877);
+        $this->expectExceptionMessage('jrid must not be overriden for table with handle "some handle", original jrid is "42", new jrid id "21"');
+
+        $dataset = [
+            'jrid' => 21,
+            'field1' => 'new content of field 1',
+            'field2' => 'new content of field 2',
+        ];
+        $this->subject->setDataset($dataset);
+    }
+}