+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Extensible formatters for table columns in content element (#10)
-- Selectable alignment for table columns in content element (#11)
+- Select alignment for table columns in content element (#11)
+- PSR-14 event to adjust or reject a dataset on synchronisation (#13)
 
 ### Fixed
 - Domain model Table::setHandle() sets the correct property