+++ Documentation/changelog.rst
@@ -16,7 +16,8 @@ Added
 
 
 * Extensible formatters for table columns in content element (#10)
-* Selectable alignment for table columns in content element (#11)
+* Select alignment for table columns in content element (#11)
+* PSR-14 event to adjust or reject a dataset on synchronisation (#13)
 
 Fixed
 ^^^^^